Update on "Remove Tensor.h, TensorMethods.h from src/core."
This is a major source of merge conflicts, and AFAICT isn't necessary anymore (it may have been necessary for some mobile build stuff in the past).

This is a commandeer of #25031

Differential Revision: [D17687345](https://our.internmc.facebook.com/intern/diff/D17687345)

[ghstack-poisoned]
