+++ aten/src/ATen/core/CMakeLists.txt
@@ -1,18 +1,29 @@
 # This file solely exists to let Caffe2 Android build get at the list
 # of core files without having to trundle through all of ATen's CMakeLists.txt
 
+FILE(GLOB_RECURSE ATen_CORE_HEADERS "*.h")
+FILE(GLOB_RECURSE ATen_CORE_SRCS "*.cpp")
+FILE(GLOB_RECURSE ATen_CORE_TEST_SRCS "*_test.cpp")
+EXCLUDE(ATen_CORE_SRCS "${ATen_CORE_SRCS}" ${ATen_CORE_TEST_SRCS})
+
 # Add files needed from jit folders
 LIST(APPEND ATen_CORE_HEADERS
-    ${CMAKE_CURRENT_SOURCE_DIR}/context_base.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/grad_mode.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/ATenGeneral.h
-)
-LIST(APPEND ATen_CORE_HEADERS
-#   ${core_generated_h}
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/source_range.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/function_schema_parser.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/lexer.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/strtod.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/parse_string_literal.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/schema_type_parser.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/error_report.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/tree.h
 )
 LIST(APPEND ATen_CORE_SRCS
-    ${CMAKE_CURRENT_SOURCE_DIR}/context_base.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/grad_mode.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/error_report.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/function_schema_parser.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/lexer.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/strtod.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/script/schema_type_parser.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../../../torch/csrc/jit/source_range.cpp
 )
 
 # Pass to parent
