+++ caffe2/core/tensor.h
@@ -9,10 +9,9 @@
 #include <c10/util/intrusive_ptr.h>
 #if !defined(CAFFE2_IS_XPLAT_BUILD) && !defined(C10_MOBILE)
 #include "ATen/core/Tensor.h"
+#include <ATen/core/grad_mode.h>
 #endif
 #include <c10/core/TensorOptions.h>
-
-#include <ATen/core/grad_mode.h>
 
 namespace caffe2 {
 
@@ -197,7 +196,9 @@ class CAFFE2_API Tensor final {
    */
   void CopyFrom(const Tensor& src, bool async = false) {
     // TODO: only check `!impl_->requires_grad()` after Variable and Tensor are merged
+#if !defined(CAFFE2_IS_XPLAT_BUILD) && !defined(C10_MOBILE)
     AT_ASSERT(!impl_->is_variable() || !(impl_->requires_grad() && at::GradMode::is_enabled()));
+#endif
     AT_ASSERTM(
         src.impl_->is_contiguous(),
         "Right now only copy of contiguous source Tensor is supported.");
